fsverity: always use bitmap to track verified status
The bitmap is used to track verified status of the Merkle tree
blocks which are smaller than a PAGE. Blocks which fits exactly in a
page use PageChecked() for tracking "verified" status. This patch
switches to always use bitmap to track verified status. This is
needed to move fs-verity away from page management and work only
with Merkle tree blocks.

Signed-off-by: Andrey Albershteyn <aalbersh@redhat.com>